split 3D APIs into separate files in /_api/three_d/
